requirements.txt
Flask
flask_wtf
psycopg2
Flask-Login
requests
flask-bcrypt
python-dotenv